rename module to match pip spec
