+++ scodec/__main__.py
