+++ scodec/__version__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+VERSION = (0, 0, 1)
+
+__version__ = '.'.join(map(str, VERSION))
