+++ scodec/codec/n3.py
@@ -10,15 +10,13 @@ Dependancies
 ------------
 ```
 from typing import List, Tuple
-
-from codec.scodec import SpatialCodec
+from scodec.codec.base import SpatialCode
 ```
 Copyright © 2021 LEAP. All Rights Reserved.
 """
 
 from typing import List, Tuple
-
-from codec.scodec import SpatialCodec
+from scodec.codec.base import SpatialCodec
 
 
 class N3(SpatialCodec):
@@ -53,7 +51,7 @@ class N3(SpatialCodec):
         bits = [bitstream >> i & 0x1 for i in range(self.block_size)]
         # generate index by encoding each set bit sequentially
         stream = list(filter(None, [self.encode(i) if b else None for i, b in enumerate(bits)]))
-        self.log.debug("stream: %s", stream)
+        self.log.info("stream: %s", stream)
         if mpl: self.render(stream)
         return stream
 
@@ -72,9 +70,9 @@ class N3(SpatialCodec):
         self.log.debug("Expected byte size: %s", block)
         for coor in stream:
             bitstream |= self.decode(coor)
-        self.log.debug("decoded bitstream: %s", bin(bitstream))
+        self.log.info("decoded bitstream: %s", bin(bitstream))
         bytestream = bitstream.to_bytes(block, byteorder="big", signed=False)
-        self.log.debug("bytestream: %s", bytestream)
+        self.log.info("bytestream: %s", bytestream)
         return bytestream
 
     def decode(self, coor: Tuple[int, int, int]) -> int:
